# frontend/api_doc.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Movie Review System API",
    "description": "Backend API for Movie Review Website",
    "version": "v1.0"
  },
  "servers": [
    {
      "url": "http://localhost:9090",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/analytics/trending": {
      "get": {
        "tags": [
          "数据分析与榜单"
        ],
        "summary": "获取趋势榜单",
        "description": "获取今日、本周、本月的热门电影。底层基于 Hive/Spark 计算。",
        "operationId": "getTrendingMovies",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "description": "周期类型: DAILY, WEEKLY, MONTHLY",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultListMovie"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Movie": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "电影ID",
            "format": "int64",
            "example": 1292052
          },
          "name": {
            "type": "string",
            "description": "电影名称",
            "example": "肖申克的救赎"
          },
          "alias": {
            "type": "string",
            "description": "电影别名",
            "example": "The Shawshank Redemption / 月黑高飞(港) / 刺激1995(台)"
          },
          "actors": {
            "type": "array",
            "description": "演员列表 (包含name和role等信息)",
            "example": [
              {
                "name": "蒂姆·罗宾斯",
                "id": "12345"
              }
            ],
            "items": {
              "type": "object",
              "additionalProperties": {
                "type": "object",
                "description": "演员列表 (包含name和role等信息)",
                "example": [
                  {
                    "name": "蒂姆·罗宾斯",
                    "id": "12345"
                  }
                ]
              },
              "description": "演员列表 (包含name和role等信息)",
              "example": [
                {
                  "name": "蒂姆·罗宾斯",
                  "id": "12345"
                }
              ]
            }
          },
          "cover": {
            "type": "string",
            "description": "封面图片路径 (自动拼接域名)",
            "example": "http://localhost:8080/images/p480747492.jpg"
          },
          "directors": {
            "type": "array",
            "description": "导演列表",
            "example": [
              {
                "name": "弗兰克·德拉邦特",
                "id": "15615"
              }
            ],
            "items": {
              "type": "object",
              "additionalProperties": {
                "type": "object",
                "description": "导演列表",
                "example": [
                  {
                    "name": "弗兰克·德拉邦特",
                    "id": "15615"
                  }
                ]
              },
              "description": "导演列表",
              "example": [
                {
                  "name": "弗兰克·德拉邦特",
                  "id": "15615"
                }
              ]
            }
          },
          "score": {
            "type": "number",
            "description": "本站评分 (0-10)",
            "format": "double",
            "example": 9.7
          },
          "doubanScore": {
            "type": "number",
            "description": "豆瓣评分 (0-10)",
            "format": "double",
            "example": 9.7
          },
          "votes": {
            "type": "integer",
            "description": "本站评分人数",
            "format": "int32",
            "example": 2956885
          },
          "doubanVotes": {
            "type": "integer",
            "description": "豆瓣评分人数",
            "format": "int32",
            "example": 2956885
          },
          "genres": {
            "type": "string",
            "description": "电影类型 (逗号分隔)",
            "example": "犯罪,剧情"
          },
          "imdbId": {
            "type": "string",
            "description": "IMDB ID",
            "example": "tt0111161"
          },
          "languages": {
            "type": "string",
            "description": "语言",
            "example": "英语"
          },
          "mins": {
            "type": "string",
            "description": "片长",
            "example": "142分钟"
          },
          "regions": {
            "type": "string",
            "description": "制片国家/地区",
            "example": "美国"
          },
          "releaseDate": {
            "type": "string",
            "description": "上映日期",
            "example": "1994-09-10(多伦多电影节)"
          },
          "storyline": {
            "type": "string",
            "description": "剧情简介",
            "example": "20世纪40年代末，小有成就的青年银行家安迪..."
          },
          "year": {
            "type": "integer",
            "description": "上映年份",
            "format": "int32",
            "example": 1994
          },
          "writers": {
            "type": "array",
            "description": "编剧列表",
            "example": [
              {
                "name": "斯蒂芬·金",
                "id": "12345"
              }
            ],
            "items": {
              "type": "object",
              "additionalProperties": {
                "type": "object",
                "description": "编剧列表",
                "example": [
                  {
                    "name": "斯蒂芬·金",
                    "id": "12345"
                  }
                ]
              },
              "description": "编剧列表",
              "example": [
                {
                  "name": "斯蒂芬·金",
                  "id": "12345"
                }
              ]
            }
          }
        },
        "description": "电影详细信息实体"
      },
...
      "ResultListMovie": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码 (200表示成功，其他为错误码)",
            "format": "int32",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "example": "Success"
          },
          "data": {
            "type": "array",
            "description": "业务数据载荷",
            "items": {
              "$ref": "#/components/schemas/Movie"
            }
          }
        },
        "description": "统一 API 响应结构"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}